🔧 chore(base.json): update compiler options in base.json 🔧 chore(cdk.json): create cdk.json and extend base.json 🔧 chore(common.json): create common.json and extend base.json 🔧 chore(package.json): add cdk.json, common.json, and react.json to files list 🔧 chore(react.json): create react.json and extend base.json
The changes in the base.json file include updating the compiler options to use uppercase values for target, module, and moduleResolution. The outDir is changed to "dist" and some strict options are added. The exclude property is also added to exclude certain directories from compilation.

The cdk.json file is created and extends the base.json file. It adds the lib property with values "DOM", "DOM.Iterable", and "ESNext". The ts-node section is also added with module set to "CommonJS".

The common.json file is created and extends the base.json file. It adds the module property with the value "CommonJS" and changes the outDir to "dist".

The package.json file is updated to include the newly created cdk.json, common.json, and react.json files in the files list.

The react.json file is created and extends the base.json file. It adds the allowJs property with the value true, sets the baseUrl to "./src", and adds some lib options. The include property is also updated to include TypeScript and TypeScript React files.